% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_lag.R
\name{compute_time_lag}
\alias{compute_time_lag}
\title{Compute Time Lag in Detection of Turning Points}
\usage{
compute_time_lag(
  data,
  peaks = nber_tp_m[, "Peak"],
  troughs = nber_tp_m[, "Trough"],
  frequency = 12,
  type = c("first_detection", "no_revisions"),
  tp_limit = 3,
  n_ahead_max = NULL
)
}
\arguments{
\item{data}{the data.}

\item{peaks, troughs}{the reference dates of peaks and troughs.}

\item{frequency}{the annual frequency of the series.}

\item{type}{method used to devine the time lag.}

\item{tp_limit}{maximal number of periods between the detected turning-point and the reference dates.}

\item{n_ahead_max}{maximal number of periods after the final detected turning point used to compute the time lag (when \code{type ="no_revisions"}).}
}
\description{
Compute Time Lag in Detection of Turning Points
}
