% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gg_functions.R
\name{ggMultisave}
\alias{ggMultisave}
\title{Save a ggplot with multiple extensions}
\usage{
ggMultisave(filename, out = c("pdf", "jpg", "svg"), ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{out}{extensions used for the export.}

\item{...}{other arguments used in \link[ggplot2:ggsave]{ggplot2::ggsave}.}
}
\description{
Save a ggplot with multiple extensions
}
