% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{revisions_stats}
\alias{revisions_stats}
\alias{first_est_revisions}
\alias{consecutive_est_revisions}
\title{Compute revisions statistics}
\usage{
first_est_revisions(x)

consecutive_est_revisions(x)
}
\arguments{
\item{x}{list of data with all the estimations.}
}
\description{
\code{first_est_revisions()} computes relative revisions to the last estimate and
\code{consecutive_est_revisions()} computes relative revisions to the next estimate.
}
