% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{simulated_tci}
\alias{simulated_tci}
\alias{cycle}
\alias{trend}
\alias{irregular}
\title{Simulate Trend-Cycle Component}
\usage{
simulated_tci(t, rho, lambda, sigma_nu, sigma_e, start = 0)

cycle(t, rho, lambda)

trend(n = 1, sigma_nu = 1, start = 0)

irregular(n = 1, sigma_e = 1)
}
\arguments{
\item{t, rho, lambda, sigma_nu, sigma_e, start, n}{parameters, see details.}
}
\description{
Simulate Trend-Cycle Component
}
\details{
\loadmathjax
The following formula is used for the cycle
\mjsdeqn{
C_t = \rho [
\cos (2 \pi t / \lambda) +
\sin (2 \pi t / \lambda)
]
}
The trend component is assumed to be a random walk with drift:
\mjsdeqn{
T_t = T_{t-1} + \nu_t\quad \nu_t \sim \mathcal{N}(0, \sigma_\nu^2)
}
The irregular component is white noise with mean zero and variance \mjseqn{\sigma_e^2}:
\mjsdeqn{
I_t = e_t\text{ with }
e_t \sim \mathcal{N}(0, \sigma_e^2)
}
}
