% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{summary_revisions}
\alias{summary_revisions}
\title{Summary statistics on revisions}
\usage{
summary_revisions(
  data,
  peaks = nber_tp_m[, "Peak"],
  troughs = nber_tp_m[, "Trough"],
  annual_frequency_tp = 2
)
}
\arguments{
\item{data}{the data.}

\item{peaks, troughs}{the reference dates of peaks and troughs.}

\item{annual_frequency_tp}{number of of observations per unit of time used
before and after the turning-points used to compute revision statistics on turning-points.
For example, \code{annual_frequency_tp = 2} means that, with monthly data, 12/2=6 months are
used before and after the turning-points.}
}
\description{
Computes summary statistics on revisions, on the the overall data and
around turning-points.
}
