% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turning_points.R
\name{turning_points}
\alias{turning_points}
\alias{upturn}
\alias{downturn}
\title{Detect turning points in a time series}
\usage{
turning_points(x, start = NULL, end = NULL, digits = 6)

upturn(x, start = NULL, end = NULL, digits = 6)

downturn(x, start = NULL, end = NULL, digits = 6)
}
\arguments{
\item{x}{the input time series}

\item{start, end}{the interval where to find turning points}
}
\description{
Detect turning points in a time series
}
\details{
Zellner, Hong, et Min (1991) definition is used \eqn{k=3}, \eqn{m=1}:
\itemize{
\item we have an upturn at date \eqn{t} when
\deqn{
y_{t-k}\geq\cdots\geq y_{t-1}<y_t\leq y_{t+1}\leq\cdots y_{t+m}
}
\item we have a downturn at date \eqn{t} when
}

\deqn{
y_{t-k}\leq\cdots\leq y_{t-1}>y_t\geq y_{t+1}\geq\cdots y_{t+m}
}
}
