% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x11_utilities.R
\name{x11_utilities}
\alias{x11_utilities}
\alias{calcICR}
\alias{calAbsMeanVariations}
\alias{selectFilter}
\alias{selectFilter.default}
\alias{selectFilter.ts}
\title{X11 utility function}
\usage{
calcICR(x)

calAbsMeanVariations(x, nlags = 1, mul = FALSE)

selectFilter(x, ...)

\method{selectFilter}{default}(x, ..., freq)

\method{selectFilter}{ts}(x, ...)
}
\arguments{
\item{x}{the input time series}

\item{nlags, mul}{parameters}

\item{...}{unused argument}

\item{freq}{the frequency.}
}
\description{
X11 utility function
}
