% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredmd.R
\name{fredmd}
\alias{fredmd}
\title{Import fredmd database}
\usage{
fredmd(
  file = "",
  date_start = NULL,
  date_end = NULL,
  transform = TRUE,
  log = FALSE
)
}
\arguments{
\item{file}{path to a file.}

\item{date_start, date_end}{optional date start and end.}

\item{transform}{boolean indicating if the data should be transformed.}

\item{log}{boolean indicating if only the log transformation should be used (\code{transform} must be equal to \code{FALSE}).}
}
\description{
Import fredmd database
}
